+++ NEWS.md
@@ -3,6 +3,10 @@ title: "Release notes for `PFUAggDatabase`"
 output: html_document
 ---
 
+* Now testing for primary, final, and useful
+  aggregations that should be equal.
+  An error is thrown if they are now.
+
 
 # PFUAggDatabase 0.1.2 (2022-06-06) [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.8011597.svg)](https://doi.org/10.5281/zenodo.8011597)