+++ NEWS.md
@@ -3,9 +3,17 @@ title: "Release notes for `PFUAggDatabase`"
 output: html_document
 ---
 
+Cite all releases with doi [10.5281/zenodo.6409759](https://doi.org/10.5281/zenodo.6409759), 
+which always resolves to the latest release.
+
+* Fixed a bug where U_eiou wasn't in argument list 
+  for a call to `Recca::finaldemand_aggregates()`.
 * Added a new vignette that shows how to 
   read data from pins and make a graph of 
   country efficiencies coloured by continent.
+* New function `rename_prime_cols()` deletes original matrices and 
+  renames "*_prime" to original names.
+  Used in the chopping process.
 * Now testing for primary, final, and useful
   aggregations that should be equal.
   An nicely formatted error is thrown if they are not.